feat: structural refactor & secrets caching
