+++ infisical_sdk/resources/auth_methods/__init__.py
@@ -1,0 +1,2 @@
+from .aws_auth import AWSAuth
+from .universal_auth import UniversalAuth
