+++ sink/README.md
@@ -1,0 +1,3 @@
+**Developer note:**
+
+To run the sink files, you must first run `pip install -e .` from inside the project root, to install the SDK in development mode.
